取平均值

+++ src/com/example/wifiscan2/AP.java
@@ -8,6 +8,6 @@ package com.example.wifiscan2;
 public class AP {
 	String SSID;
 	String BSSID;
-	int level;
+	double level;
 }
